gzap: scanner rewrite
It now uses integer location coordinates and captures them in WorldThingSpawned,
so no more inconsistencies due to gravity, Z-correction, elevators, etc -- the
coordinates in the map file are what's recorded.

+++ gzap/apworld/gzdoom/model/DoomLocation.py
@@ -16,9 +16,9 @@ class DoomPosition(NamedTuple):
     """
     map: str
     virtual: bool  # True if this doesn't actually exist in the world and coords are meaningless
-    x: float
-    y: float
-    z: float
+    x: int
+    y: int
+    z: int
 
 
 class DoomLocation:
